Made 2 different LinkedLists one with raw pointers and one with smart pointers
